+++ wxc_sdk/as_api.py
@@ -53,8 +53,8 @@ __all__ = ['AsAccessCodesApi', 'AsAdminAuditEventsApi', 'AsAgentCallerIdApi', 'A
            'AsOperatingModesApi', 'AsOrgEmergencyServicesApi', 'AsOrgMSTeamsSettingApi',
            'AsOrganisationAccessCodesApi', 'AsOrganisationVoicemailSettingsAPI', 'AsOrganizationApi',
            'AsOrganizationContactsApi', 'AsOutgoingPermissionsApi', 'AsPSTNApi', 'AsPagingApi', 'AsPeopleApi',
-           'AsPersonForwardingApi', 'AsPersonSettingsApi', 'AsPersonSettingsApiChild', 'AsPlayListApi',
-           'AsPreferredAnswerApi', 'AsPremisePstnApi', 'AsPriorityAlertApi', 'AsPrivacyApi',
+           'AsPersonForwardingApi', 'AsPersonSettingsApi', 'AsPersonSettingsApiChild', 'AsPersonalAssistantApi',
+           'AsPlayListApi', 'AsPreferredAnswerApi', 'AsPremisePstnApi', 'AsPriorityAlertApi', 'AsPrivacyApi',
            'AsPrivateNetworkConnectApi', 'AsPushToTalkApi', 'AsRebuildPhonesJobsApi', 'AsReceptionistApi',
            'AsReceptionistContactsDirectoryApi', 'AsRecordingsApi', 'AsReportsApi', 'AsRestSession', 'AsRolesApi',
            'AsRoomTabsApi', 'AsRoomsApi', 'AsRouteGroupApi', 'AsRouteListApi', 'AsSCIM2BulkApi', 'AsSCIM2GroupsApi',
@@ -9175,7 +9175,7 @@ class AsTransferNumbersApi(AsPersonSettingsApiChild):
 
     async def configure(self, entity_id: str, settings: AutoTransferNumbers, org_id: str = None):
         """
-        Modify Transfer Numbers Settings for a Place.
+        Modify Transfer Numbers Settings for an entity.
 
         When calling a specific call type, this entity will be automatically transferred to another number.
         The person assigned the Auto Transfer Number can then approve the call and send it through or reject the
@@ -9362,6 +9362,94 @@ class AsPersonForwardingApi(AsPersonSettingsApiChild):
         params = org_id and {'orgId': org_id} or None
         data = forwarding.update()
         await self.put(ep, params=params, json=data)
+
+
+class AsPersonalAssistantApi(AsApiChild, base=''):
+    """
+    API for personal assistant settings
+
+    """
+
+    async def get(self, person_id: str, org_id: str = None) -> PersonalAssistant:
+        """
+        Get Personal Assistant
+
+        Retrieve Personal Assistant details for a specific user.
+
+        Personal Assistant is used to manage a user's incoming calls when they are away.
+
+        Retrieving Personal Assistant details requires a full, user, or read-only administrator auth token with a scope
+        of `spark-admin:telephony_config_read`.
+
+        :param person_id: Unique identifier for the person.
+        :type person_id: str
+        :param org_id: Get Personal Assistant details for the organization.
+        :type org_id: str
+        :rtype: :class:`PersonalAssistant`
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        url = self.ep(f'telephony/config/people/{person_id}/features/personalAssistant')
+        data = await super().get(url, params=params)
+        r = PersonalAssistant.model_validate(data)
+        return r
+
+    async def update(self, person_id: str, enabled: bool = None,
+               presence: PersonalAssistantPresence = None,
+               until_date_time: Union[str, datetime] = None, transfer_enabled: bool = None, transfer_number: str = None,
+               alerting: PersonalAssistantAlerting = None,
+               alert_me_first_number_of_rings: int = None, org_id: str = None):
+        """
+        Update Personal Assistant
+
+        Update Personal Assistant details for a specific user.
+
+        Personal Assistant is used to manage a user's incoming calls when they are away.
+
+        Updating Personal Assistant details requires a full or user administrator auth token with a scope of
+        `spark-admin:telephony_config_write`.
+
+        :param person_id: Unique identifier for the person.
+        :type person_id: str
+        :param enabled: Toggles feature.
+        :type enabled: bool
+        :param presence: Person's availability.
+        :type presence: PersonalAssistantPresence
+        :param until_date_time: The date until which the personal assistant is active.
+        :type until_date_time: Union[str, datetime]
+        :param transfer_enabled: Toggle the option to transfer to another number.
+        :type transfer_enabled: bool
+        :param transfer_number: Number to transfer to.
+        :type transfer_number: str
+        :param alerting: Alert type.
+        :type alerting: PersonalAssistantAlerting
+        :param alert_me_first_number_of_rings: Number of rings for alert type: ALERT_ME_FIRST; available range is 0-20.
+        :type alert_me_first_number_of_rings: int
+        :param org_id: Update Personal Assistant details for the organization.
+        :type org_id: str
+        :rtype: None
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        body = dict()
+        if enabled is not None:
+            body['enabled'] = enabled
+        if presence is not None:
+            body['presence'] = enum_str(presence)
+        if until_date_time is not None:
+            body['untilDateTime'] = until_date_time
+        if transfer_enabled is not None:
+            body['transferEnabled'] = transfer_enabled
+        if transfer_number is not None:
+            body['transferNumber'] = transfer_number
+        if alerting is not None:
+            body['alerting'] = enum_str(alerting)
+        if alert_me_first_number_of_rings is not None:
+            body['alertMeFirstNumberOfRings'] = alert_me_first_number_of_rings
+        url = self.ep(f'telephony/config/people/{person_id}/features/personalAssistant')
+        await super().put(url, params=params, json=body)
 
 
 class AsPreferredAnswerApi(AsApiChild, base='telephony/config/people'):
@@ -10797,6 +10885,8 @@ class AsPersonSettingsApi(AsApiChild, base='people'):
     permissions_in: AsIncomingPermissionsApi
     #: Person's Outgoing Calling Permissions Settings
     permissions_out: AsOutgoingPermissionsApi
+    #: Personal Assistant Settings
+    personal_assistant: AsPersonalAssistantApi
     #: Preferred answer endpoint settings
     preferred_answer: AsPreferredAnswerApi
     #: Person's Privacy Settings
@@ -10846,6 +10936,7 @@ class AsPersonSettingsApi(AsApiChild, base='people'):
         self.numbers = AsNumbersApi(session=session)
         self.permissions_in = AsIncomingPermissionsApi(session=session)
         self.permissions_out = AsOutgoingPermissionsApi(session=session)
+        self.personal_assistant = AsPersonalAssistantApi(session=session)
         self.preferred_answer = AsPreferredAnswerApi(session=session)
         self.privacy = AsPrivacyApi(session=session)
         self.push_to_talk = AsPushToTalkApi(session=session)